Move equity chart to `src/gui`.
