+++ classMainWindow.py
@@ -19,7 +19,7 @@ from copy import deepcopy
 from src.gui.dialog_box import ConnectWindow, OptionsWindow, FilterWindow, AboutWindow, ExportWindow
 from src.gui.widgets import CustomLabel, CustomLineEdit, CustomDockWidget
 
-from classEquityChart import EquityChart
+from src.gui.classEquityChart import EquityChart
 
 import classRestCom
 import classLsEvent
